[network][noise] always send timestamp in prologue
We currently send a timestamp as part of the Noise handshake's first message payload
only if we are in the mutually authenticated network.

This currently make the logic more complex, which can have well-understood consequences...
Especially after a refactor.
So right now, a server in the full node network still waits for a message large enough to contain a payload.
This is bad, this commit fixes this and simplifies the protocol.