[storage] Replace the Box<dyn Storage> runtime shim by a static enum
We use an open trait hierarchy for `CryptoStorage + KVStorage`, which we represent with a(n emulated) trait-alias `Storage`, which carries signing and verification APIs through the `CryptoStorage` component of the trait alias.
We are also keen to be able to instantiate the implementations of this hierarchy through a factory method taking a configuration as an input, speficically a `SecureBackend`, which is a closed sum type (an `enum`). Hence this conversion must return trait objects (specifically `Box<dyn Storage>`), and crucially relies on this dispatch.

This has two consequences:
- since trait objects cannot include generic methods, the sign and verify methods of `CryptoStorage` and all its dependents (`CryptoHash` in particular) would have to take `v: &dyn Cryptohash` arguments instead of `v: CryptoHash`,
- this reliance on dynamic dispatch is slow: I measure about 2000 microseconds spent on this dynamic dispatch, against 100 microseconds for actually verifying a signature afterwards.

The present PR re-establishes static dispatch through a mirror enum, which instantiates a closed sum type over the inhabitants of the `KVStorage + CryptoStorage` hierarchy. That's what the conversion from `SecureBackend` returns. It should offer better performance, and allow `CryptoStorage` to offer sign & verify APis that take a `x: CryptoHash + Storage`.

Details:
- Don't worry about the `BoxedStorage` name, I'll modify it programmatically to `Storage` (o.b.o.), it's meant to help review temproraily.
- The enum and all its contents can thankfully all be generated automatically, which should help with the maintenance of the boilerplate in `libra_secure_storage::storage`. A good starting point to do that is the [enum_dispatch](https://gitlab.com/antonok/enum_dispatch) crate, which lacks two things in order to work for us as is:
  + support for implementation of multiple traits,
  + support for a `Box<dyn T>` parameter for an enum variant, because ...
- The KVStorage + CryptoStorage hierarchy contains a recursive variant (NamespacedStorage), which obviously complicates the enum definition a bit (the conversion from SecureBackend would not know how to return a `BoxedStorage<T>` without knowing `T`), and hence requires me to use a (slow) `dyn Trait` in the `BoxedStorage` destination enum.
- I've removed the bespoke trait alias emulation performed to equate `Storage` with `KVStorage + CryptoStorage` because I expect macro-powered generation `impl`s of an enum closing over a trait hierearchy to be simpler if the traits one must implement on the enum are direct requirements in client code.