[LEC] build the authenticated link between LSR and LEC on LEC side
This PR setup the authenticated link between LSR and LEC. The basic rationale behind
this PR is the architecture of TCB. LEC is in zone 2, safety-rules is in zone 1 and
consensus is in zone 3. Because LSR doesn't trust consensus, the computation result
from LEC cached in consensus should be authenticated by LSR to verify the integrity
to avoid any tampering from any third party including consensus.

To do this, we add a key pair `EXECUTION_KEY` into vault whose private key is only
accessible by LEC. Any state_compute_result returned from LEC will carry a signature
of a proposal associated with this block. When this block needs to be voted, the generated
vote proposal will be verified by LSR service using the public key.

This commit build the signing side on LEC and corresponding changes in consensus and configs.