[network] properly configure server only -- sort of
This PR makes the gross assumption that a peer without seed peers is a
server and one with seed peers is a client. It then configures the
network to build accordingly.

Inside connectivity manager, it merges the seed peers and network peers,
so that it can make and retain outgoing connections.

As a result, the smoke_test now passes and we have connectivity our VFN
working as expected (minus services, which are not yet tested).

In the next PR, we'll need to clean this up, because we'll add another
service that will actually use full node onchain discovery.