[move core][DRAFT] Added module to query genesis status.
This file implements a state variable to track the status of genesis.
It is in a separate leave module, because it will be used by lots of
other modules to check whether functions are being called during the
genesis process or not. Putting it in Genesis would introduce a ton
of dependencies, but using GenesisState will not.

In the Genesis module, the very first thing `initialize` should
to is to call begin_initialize().  The last thing it should do
is call end_initialized().

The implementation is a state machine with three states: before,
during, and after genesis. `begin_genesis` advances the state
machine from before to during, and `end_genesis` advances it
from during to after. Each of these transitions is irreversible.

Functions in other modules, especially initialization functions, can
call predicates is_during_genesis(): bool and is_after_genesis: bool
return if genesis is underway or complete, respectively.