[config] Provide access to default configs
Note these configs are "default" in the context that a test can override
some of the per-config specifics (e.g., which secure store or which
private key), but still dictate default network layout and other system
aspects. They will be useful for moving forward with smoke tests
involving configs. This will also let us ensure that our documented
defaults are correct and work.