[backup] Refactor backup service
Updated the way we deal with iterator based endpoints
(/state_snapshot/<version>), now the iterator
can carry a reference to something else, as long as we pack the referee
into a same closure (alternatively we can let the iter itself carry the
referee, which requires a self-referencing structure, which in turn
requires unsafe code). This is preparation for the coming epoch ending
ledger info iter and transaction iterator, both of which involve a
reference to the DB.

Also restructured the crate a little, to separate out utils from the
handlers mod.