[language] Clean up Move VM API
This removes `cache_module()` from the runtime and VM API which is
a good thing. That entry point had a usage that was testing and
convenience only.
We can replace it everywhere with proper changes to the data
view and with a publish/execute model.
The tricky usage is, as expected, in the genesis generator
but the change is very contained and could be argued the
right thing.