[move-prover] Implemented three initial specification checks.
[Cleaned up specification check setup
Introduced two additional enum types for `ConditionKind`.
`RequiresSmokeTest` which are expressions assumed at the beginning
of the top level `_verify` function and `EnsuresSmokeTest` which
is asserted at the end of the top-level `_verify` function.

Using these, I added a test to avoid false positive `aborts_if true;`
checks and a fix:

The smoke test option should only be used by the module level
spec block and not individual functions (to be added) otherwise the
pre/post condition errors from callees will mix with the smoke
test errors.

[Constant global resource spec check]

Added instrumentation to check for unused fields.

Added a wrapper function for each specification check
smoke test because negative errors requires that all
the errors are reported. This can be further optimized but
currently a new wrapper function is generated for each
expression of kind EnsuresSmokeTest.

The issue of using one function is that if multiple negative specification
check are in a single function, Boogie does not necessarily return
all of these errors. Then it is not clear whether all the
negative errors were violated.

[Extended ast to have constructors]

Added constructor helper functions inside the spec-lang ast.

When building ast nodes for the specification checking (SC) tool,
the expression has no notion of location in a file. So building
these constructors use a dummy location.

Fixed field specification check to filter out resources that aren't
moved to the top level of the global store.

[Constant sub-expression spec check]

Implemented constant sub-expression check as a specification check.
Also rewrote backend to allow for multiple assertions for a single function.

[Writeref specification check]

Implemented the writeref specification check.
For each WriteRef inside the (byte) code of a function,
generate a new list of bytecodes where the WriteRef's value is
replaced with a $DefaultValue(). If the user has not specified
the mutation in the specifications, the check will raise an
error.

NOTE: Instead of using a $DefaultValue(), later this is replaced with
a fresh unconstrained variable.

[Fixed specification check for constant subexpressions]

[Fixed specification check bugs]

1. Added initial implemention of printing for the spec-lang ast.
   NOTE: Removed during merging.
2. Fixed bug in constant field checker to check for borrow globals
   on a given resource.
3. Refactored so verification is disjoint from smoke testing entirely.
   Now we check if any smoke test option is declared true, we ignore
   all specifications and inject specification checks according to
   the given pragmas.
4. When in specification mode, only generate the indirect functions.
5. Rebased with master

[Mutated arg spec check]

Replaced separate rewritten_spec (Vec<Spec>) and
rewritten_code (Vec<Vec<Bytecode>>) with a SpecCheck
struct that holds the specification check as a Vec<Spec>
and an optional Vec<Bytecode> if the function code is mutated.

Added annotation to location for discerning between locations in the
file and synthetic locations from generating specification checks.
This is used to discern between multiple negative errors for the same
specification check.

Implemented a specification check that mutates the values of return
values of a procedure. This operates at the Boogie level, which means
it includes both mutatable references (since these are included in the
return signature after the return values of a function) and return
values of the Move function being verified.

[Updated specification check tests]

Added a specification check for `mutated_args_test` and updated others.
Currently, specification checks require declaring `pragma verify = false`.