[move-prover] Constant global resource spec check
Added instrumentation to check for unused fields.

Added a wrapper function for each specification check
smoke test because negative errors requires that all
the errors are reported. This can be further optimized but
currently a new wrapper function is generated for each
expression of kind EnsuresSmokeTest.

The issue of using one function is that if multiple negative specification
check are in a single function, Boogie does not necessarily return
all of these errors. Then it is not clear whether all the
negative errors were violated.