[move-prover] Add missing Escape.exp
I forgot Escape.exp in a PR, which cause cargo test for the Prover to fail.

This just needs a rubber stamp (assuming it gets through testing).

Note: My freshly updated "master" branch was failing in cargo test on
Libra.move.  This problem is unrelated to changes in this PR (and it
is happening in other PRs).