[move-prover] Partial global specifications of Roles.move
There is a verified specification that, for each address with a
RoleId, the role id matches the privilege stored at that same address.
This works around a Move Prover limitation by defining a pre-condtion
that the "witness" arg to the public add_privilege_to_* functions
cannot be one of the role privileges defined in this module.  This
solution is not entirely satisfactory because it is not modular --
correctness depends on verifying that the precondition is met in every
module that uses Roles, which we probably don't want to assume in a
system with third-party Move programs.

In addition, there is a specification that, once a Role is assigned,
it cannot change.

There is also a preliminary attempt to specify that the "is_extracted"
flag for a a privilege is set if and only if there exists a capability
for that privilege.  I'm not sure it's correct, and the verifier gets
stuck on it, so 3 out of 4 properties are commented out for now.