[move framework][move-prover][DRAFT] Refactor coins for better verification
Previous, Genesis called Coin1::initialize, which returned mint and
burn capabilities.  Genesis then called
LibraAccount::create_treasury_compliance_account with these, and that
function called Libra::publish_mint_capability with it as an argument.

This code bypasses genesis. The capabilities are published directly in
Coin1.  The main advantage is that this will make it possible to
verify that only the treasury compliance account can mint Coin1
based only on the code in the Coin1 module and its dependencies.
This is a step in trying to make it easier to prove properties such
as only the specified account can mint the coin.

Similarly, Coin2 for all of the above.