[language] Verifier cleanup
Here we complete the major clean up of the verifier.
There is a bit more to do to make everything more consistent but that
is not for this PR which is already massive.
VerifiedModule/Script are completely removed, given that was the general feeling in the team.
*View are removed from the verifier in favor of ad-hoc structure needed for verification.
CompiledScript is verified without converting to module. That implies that each verification pass
has 2 entry points, one for CompiledModule, the other for CompiledScript.

The common API is in binary_views.rs which provides a common access to the binary
and a unified function view.

We also made dependencies check a pass like the others, which cleaned up things quite a bit