[move-prover] Added spec for Coin1
I wanted to see how far I could get with proving that only the
treasury compliance account can mint Coin1.

Unfortunately, I was able to prove it when it's not true! The problem
is a generic function in another module that can invalidate the invariant.

In the process, I changed spec_has_..._role functions in Roles.move to
take addresses instead of signers as arguments, because it seems to be
more appropriate in specifications.  The corresponding Move functions
all take signers as arguments, which has not been a problem, so far.