[move-prover] Turning module invariants of VASP on again.
In #4870, we needed to turn off module invariants in VASP because otherwise lots of VCs in LibraAccount would not terminate. This PR tries to analyze the situation by removing some redundancy from the way how we deal with module invariants.

Consider:

```
fun is_vasp(addr): bool { exists<ChildVASP>(addr) }

schema NoChange {www
    ensures forall a: address where P(a): f(a) == old(f(a));
}

apply NoChange to *;
```

We add an `ensures` to `is_vasp` which does not make any sense, because the function is pure as can be easily seen. Z3 will indeed easily prove that this property is true.

However, every *caller* of is_vasp will get the quantifier in its VC context. Since the memory $m can have arbitray values at arbitrary addresses, Z3 may be endlessly trying to instantiate the quantifier on address terms existing in the context to disprove the quantifier. However, since the function is pure, the property is trivially true.

This PR tries to show that this theory has merit.

- All pure functions in VASP and Roles have been excluded from applying invariants as above (if this works out, it can be automated by marking functions pure via bytecode analysis)
- The representation of module invariants has been optimized in generated bytecode. We now emit a free requires instead of an assume at caller site and a requires at callee site. This is not fully done so some tests fail in corner cases.

All global properties in VASP could be turned on again, and the overall verification time of LibraAccount actually still improved. Unfortunately though, they are some functions in VASP now which do not terminate verification timely (and those react strongly on --seed settings). This need to be further investigated.