[state-synchronizer] remove Option qualifier from coordinator.waypoint
SyncCoordinator previously had an Option<Waypoint> field in order to support integration_test.rs;  this necessitated the addition of code paths that were test only and introduced unnecessary Option check code in production code paths.

This change (a) removes the option qualifier and (b) updates the tests to use default "0" values for waypoints.