[crypto] Put the fiat feature in control over whether we even load the <foo>25519-dalek packages
We use formally verified crypto (https://crates.io/crates/fiat-crypto) in maintained forks of the dalek suite of libraries ({curve, ed, x}25519-dalek).
This usage if controlled by a feature in the dependent crates, which we always use, simply turning the feature on by default.

This construction provides meaningful safety improvements against a class of arithmetic bugs in Libra's dependent libraries. We hope to merge it upstream in the future.

In some contexts (e.g. vendored dependencies), where it is difficult to load several versions of packages, we would like to not even download those forked packages, and rather use the underlying vanilla projects from the dalek suite of libraries.

This PR offers this opportunity by putting a feature in control of the choice of dependency (fork vs. vanilla dependency), rather than whether the formally verified backend is activated.

The behavior is not changed: the formally verified backend is still activated by default, and its vanilla variant is still tested in CI.