[move-prover] Writeref specification check.
Implemented the writeref specification check.
For each WriteRef inside the (byte) code of a function,
generate a new list of bytecodes where the WriteRef's value is
replaced with a $DefaultValue(). If the user has not specified
the mutation in the specifications, the check will raise an
error.

NOTE: Instead of using a $DefaultValue(), later this is replaced with
a fresh unconstrained variable.