[move-prover][libra-framework] Pragma `export_ensures` + AccountLimit specs
This PR adds some specs to AccountLimits (not complete). In the way of doing so, it became apparent that error reporting in the new translation scheme is sometimes inprecise. While the root cause for this is unclear, it appeares once all actual errors are removed from the spec/code, it works.

As a workaround, a new pragma `export_ensures` is introduced. This can be temporarily used on module level to fall back partly to the old compilation scheme (namely ensures of inlined functions are generated). This seems to fix the error problem, but is off by default because not every test is termination with it.