[config] Build VFN with proper fallback networks
There were a bunch of bugs due to implicit ordering of networks, and an
implicit number of networks.  This tries to undo that by making more
explicit based on data, and simplifying the data model.

This also ignores a couple tests because the model has changed, and in
order to move quickly I've disabled them.  Some of the previous workings
didn't work well with the newer model.