[move-prover] Mutated arg spec check.
Replaced separate rewritten_spec (Vec<Spec>) and
rewritten_code (Vec<Vec<Bytecode>>) with a SpecCheck
struct that holds the specification check as a Vec<Spec>
and an optional Vec<Bytecode> if the function code is mutated.

Added annotation to location for discerning between locations in the
file and synthetic locations from generating specification checks.
This is used to discern between multiple negative errors for the same
specification check.

Implemented a specification check that mutates the values of return
values of a procedure. This operates at the Boogie level, which means
it includes both mutatable references (since these are included in the
return signature after the return values of a function) and return
values of the Move function being verified.