[management] use get_account_transaction for validate-transaction
This will give us a rich response upon execution, no response if it has
yet to execute, and an error otherwise. Versus the previous step that
merely informed you that the transaction had been executed.

This is quite ugly in that we do not have the LCS encoded transaction
retrieval yet, but it will do for now.