[Libra Framework] Proposed handling of errors/usage of new prover features.
This PR introduces a new module `CoreErrors` which contains functions for global error codes (similar as `CoreAddresses` today). It then uses this module in some places in the framework and applies the new-style abort specifications with codes. Abort specs with codes make only sense with global errors, as they propagate through the stack. This work was started with `LibraTimestamp`, which has a revamped specification following the new methodology. The changes propagated to users of time stamp, of which there are a few.

The approach to global error codes is based on the principle that the set of codes should be small and describe the problem to end users, not the implementation details. This is assuming that debugging can be achieved by other means (replay, VM stack trace).

This work is not complete at all, and need to be continued in subsequent PRs. However, before continuing, we need to agree on the handling of global errors, so this represents a first step as a discussion basis.

Tool changes:
- Making warnings by default off. There is an option `--warn` to turn them on. This helps us for now to quiet warnings about unused schemas, which are needed for the style of spec in this PR.
- There was a bug in the computation of locations for error messages about wrong abort codes. For now, the fancy previous way which was buggy has been turned off.