[language] Detect and error on circular dependencies
Make the loader check for circular dependencies while loading.
That seems important given modules are only pushed through WriteSets
and thus never going through the verifier. If a circular dependency
was introduced, we would enter an infinite loop.
So it's a protection against an unlikely but horrible scenario.
It also fixes a problem with publishing that was loading and caching
all dependencies except the published module, though it was intended
to load nothing. Now publishing a module only verifies the dependencies
of the published module and nothing else, it also does not load any module.
That hopefully simplifies the code a bit