[language] add tests for txn validation, with fixes for gas currency
Add some missing tests for transaction validation, in both the unit
tests for the vm-validator and in the e2e verify_txn.rs tests, which
also check for consistency between validation & execution.

This includes some minor changes to validation for the gas currency
specifier. The execution phase was not checking the gas currency
specifier directly. A bad currency specifier would still cause a
failure when running the prologue, but this led to a discrepancy with
the validation phase. Besides sharing the same check in both phases,
this changes to use the CURRENCY_INFO_DOES_NOT_EXIST status code for
the case where the CurrencyInfo resource blob cannot be loaded. That
had previously been reported with a MISSING_DATA status, but aside
from being less specific, using that status with VMStatus::Error
leads to an assertion failure because it is only intended to be used
for a runtime execution error.

Closes: #5791
Closes: #6018