[Logging] Avoiding possible DoS through logs
This is adding a check that the data is coming from a somewhat
trusted or controlled network when we log the messages themselves,
to prevent a public full node to be able to DoS our log system with
super large messages.

We only log the possibly big part if we are on the Private or
Validator network.