logger: enable dynamic adjustment of both local and remote log filters
Now that we have two separate log filters, one for what is printed
locally and one for what is sent off to the remote logging service we
want to allow for dynamically adjusting both of these filters at
runtime.

This patch changes the existing `/log` route of a node's debug interface
by splitting it into `/log/filter` and `/log/remote-filter` for
configuring the local and remote filters respectively. E.g.

    # Change the local filter
    > curl http://localhost:<debug-port>/log/filter -d executor=info

    # Change the remote filter
    > curl http://localhost:<debug-port>/log/remote-filter -d debug

As before, the payload sent to these two endpoints are the same format
as the filter set with the `RUST_LOG` environment variable.

Closes: #6374