[backup] `Stream::buffered_x()`
A version of the `futures::stream::StreamExt::buffered()` that adopts a limit on concurrency.

We used to use `buffered()` to prefetch backup contents in concurrency and commit them in sequence in the main thread. But the buffer size and degree of concurrency are the same number, which is not ideal: some can't use too large a number or else we spawn too many working processes which brings in huge overhead, but sometimes when one job is slow, not every CPU core is utilized because jobs after the slow job are already done and occupying places in the buffer in vein. With the buffer size and concurrency limit separated, we can make sure the downloading is happening all the time but not context switching between them too much.