[vm] refactor validation code to allow calling it from the execution phase
This refactoring fixes a minor inconsistency in the way that WriteSet
transactions are validated during the execution phase -- the checks for the
gas currency were not done during execution. This is not a huge problem
because WriteSet transactions are not charged for gas, but it shows the
potential for inconsistencies. I'd like to avoid any potential for that in
future.

In the process of making this change, I discovered that we are charging gas
for the number of accounts that are mutated, but the VM's num_mutated_accounts
function is incorrectly including accounts that are only read from in that
count. Because of that bug, this is a breaking change. The prologue functions
are now run in a separate VM session, so with this change, no gas is charged
for the prologue functions (as was originally intended). We should still fix
that bug because it continues to affect the gas charges for executing the
transaction scripts.