[vm] refactor validation code to allow calling it from the execution phase
This refactoring fixes a minor inconsistency in the way that WriteSet
transactions are validated during the execution phase -- the checks for the
gas currency were not done during execution. This is not a huge problem
because WriteSet transactions are not charged for gas, but it shows the
potential for inconsistencies. I'd like to avoid any potential for that in
future.

In the process of making this change, I discovered that we are charging gas
for the number of accounts that are mutated, but the VM's num_mutated_accounts
function is incorrectly including accounts that are only read from in that
count. We should fix that in the future, but to preserve compatibility with
the current behavior, I had to pass around a VM session so that the gas cost
from running the prologue (due to the bug) could be included with the gas
cost from executing the transaction script.