[common] Simple token bucket rate limiter
This token bucket rate limiter is made to be generic to be able to
throttle on arbitrary tokens, with a default bucket size and fill rate
for all buckets.  The default interval in which buckets are being
refilled is 1 second.  It also provides the framework to build into the
builder ways for custom rates for specific keys, and eviction policies
of cached keys in the future.  This version is not thread safe, and must
be single threaded.