[prover] trait for compositional interproc analysis, use for packed types analysis and usage analysis
This commit introduces a `CompositionalAnalysis` trait that facilitates bottom-up interprocedural analysis. It uses the same mechanism for storing and retrieving summaries that UsageAnalysis already uses.

To demonstrate the generality of this trait, the commit splits the two analyses in usage_analysis.rs (packed types and memory usage) into seprate files and uses the `CompositionalAnalysis` trait to implement both of them.

The motivation for this is making it easier to build other summary-based analyses that are flow-sensitive (e.g., fine-grained read/write set analysis to support transaction parallelization and AOS approvals, inference of post/aborts conditions, ...). There is more work to be done first, such as allowing the summary and abstract state to be different types, supporting recursion, decreasing the boilerplate `impl trait for `, and so on. But I think this is a good start.

Note: `CompositionalAnalysis` implements a flow-sensitive propagation, which is overkill for flow-insensitive packed types and memory usage analysis. However, I measured the performance before/after conversion (benchmark: building packed types map for stdlib) and can't observe a difference. So I think it is ok to use CompositionalAnalysis everywhere, even on flow-insensitive summary-based analyses. We can always add a separate trait for flow-insensitive analyses if perf becomes a problem.
Closes: #6983