Make AccountState::get_resource_impl public again
The get_resource function used to be public: AOS is currently using it for retrieving the DD tier info for the specific currencies.
One of the recent changes closed this function, but we should still keep it available.

On a side note, we should probably define some sort of `DesignatedDealerTierInfoResource` in diem/diem, similarly to `PreburnResource`.

Closes: #6885