[rate-limiter] Add main function for throughput tests on a socket
I've added tests that test for backpressure, as well as how well the
rate limiting occurs if it's on the reader, writer, or both.  I've left
these tests ignored, because they open a socket, and need to fill the
socket buffer in order to give meaningful data which would slow
development and likely be flaky.