gha: checkout tip of PR instead of merge commit
By default the checkout action will checkout the merge commit between
the PR and the current tip of the base branch. This leads to somewhat
confusing behaviour as the tests being run in CI are not isolated to
testing just the changes introduced by a particular PR.

Instead follow the suggestion in the actions/checkout readme
https://github.com/actions/checkout#checkout-pull-request-head-commit-instead-of-merge-commit
and use the ref `${{ github.event.pull_request.head.sha }}` which will
be populated with the tip of a the PR in the case where the triggering
event is from a pull request and will be empty in the case where the
triggering event is a push to a branch which will result in the default
behaviour from the checkout action and will simply checkout the tip of
the newly pushed branch.