[network] Metric if network identity doesn't match onchain
It's possible that a user could set their network identity onchain and
then continue communicating on the network with the wrong network
identity.  This causes a strange 1 way communication, which makes it
very confusing.  The metric and error log will give us visibility into
mismatches, and make it easier for us to identify the issue.

Closes: #7217
Closes: #7669