gha: disable sccache
We're currently seeing an issue with some tests that use the trybuild
crate to run. When trybuild is compiled it looks at the `CARGO` env var
in order to determine which `cargo` binary to use during tests.
Unfortunately sccache isn't able to properly cache artifacts which
depend on environment variables at compile time, resulting in the cached
version of trybuild trying to use a nonexistent version of cargo. This
should be fixed once https://github.com/mozilla/sccache/issues/804 is
fixed.

Closes: #7241