[vm] add OUT_OF_GAS special case in VMError::into_vm_status debug_assert
A transaction may run out of gas when updating the global state, after executing
the Move code, and in that case VMError:into_vm_status is called with no location
or offset information. Ignoring the assertion (e.g., in a release build), the
code does the right thing in that case: it reports VMStatus::Error(OUT_OF_GAS)
instead of VMStatus::ExecutionFailure(OUT_OF_GAS). The ExecutionFailure status
is supposed to be used for errors that occur while executing Move code, so an
Error status seems correct. We generally do not use the same status code in
different buckets, but OUT_OF_GAS is already handled that way elsewhere.

Fixes: #7614