[DRAFT][diem-framework][move-prover]
Partially fixed some timeouts in DiemAccount.  These have been there
for awhile, but were recently exposed when we fixed some soundness
bugs in the Prover.

Problem 1: An emit was redundantly included in preburn specs,
indirectly by including PreburnWithResourceEmits both
in Diem::PreburnToEnsures and directly in the spec fun preburn.
Redundant aborts_ifs are ok, but redundant emits are not, because
the emit specs and emits in code must match 1-for-1.  I moved code
around in Diem.move to eliminate the redundancy.

Problem 2: DiemAccount::preburn calls withdraw_from, but didn't
spec the emits for withdraw_from.  I worked around this by adding
pragma emits_is_partial, because of next problem.

Problem 3: (Unresolved -- Junkil, could you please take a look?)
preburn calls withdraw_from, but when I include WithdrawFromEmits in
the spec fun preburn, it times out.  I can't see any errors in
WithDrawFromEmits, so I hope Junkil can look at this.

Problem 4: (Unresolved -- Wolfgang, could you please take a look?) I'm
not absolutely sure this is a problem, but I'm worried. In Diem.move,
UpgradePreburnEnsures has a condition in an include. The code only
does something if there exists a Preburn and if there is no
PreburnQueue; if so, it removes the Preburn and publishes a
PreburnQueue.  The condition in the ensures is exists<Preburn> and
!exists<PreburnQueue>.  However, I think that these terms refer to the
NEW values, and that the antecedant of the condition is always false,
rendering the specification trivial. The parser can't deal with "old"
in the places I want to put it.

I also removed some redundancies in the specs for a few functions in
Diem when I noticed things as I was reading them.