[move-prover] Reactivate CVC4 tests and document usage.
This reactivates a select set of cvc4 tests to be run locally (and after we add it, also nightly). This became possible after upgrading to the newest Boogie version which had the interaction issue with cvc4 fixed.

I have not really attempted to activate more tests for CVC4 in this PR, nor verified the existing ones, but only checked whether the previous fatal errors are gone.

This also adds a document at [doc/dev/cvcv4.md](https://github.com/wrwg/diem/blob/cvc4/language/move-prover/doc/dev/cvc4.md) which describes how
to select cvc4 tests, extend them, and configure the prover to use different encodings and theories if cvc4 is active. There are some TODOs in this doc which we should fill out next.

Closes: #8332