[test-infra][e2e] make sure to test until the latest DiemVersion
The original e2e tests are hardcoded to test up to version 2 while the
latest DiemVersion is at 3. This commit ensures that the e2e tests will
at least attempt to test with whatever the latest DiemVersion is.

Unfortunately, there is no upgrade path to version 3 yet so this version
is actually ignored in e2e tests. Nevertheless, it might still be good
to have the code here and once we have the upgrade path to version 3,
the e2e test executor will automatically pick it up.