[move-prover] use the new TypeUnification scheme to simulate the old behavior
The problem of checking global invariants in a generic context can be
better illustrated with this commit:

If the global invariant I talks about a generic resource, e.g., S<X, Y>,
and the implementation might also be parameterized f<X, Y>. Then, there
should be a procedure to decide whether a global invariant I should be
checked in the body of f.

                    GCD               LCM
f<u64, bool>
  I<u64, bool>  --> Match             Match
  I<u64, Y>     --> LHS               RHS
  I<X, bool>    --> LHS               RHS
  I<X, Y>       --> LHS               RHS

f<u64, Y>
  I<u64, bool>  --> RHS               LHS
  I<u64, Y>     --> Match             Match
  I<X, bool>    --> Both(u64, bool)   Both(X, Y)
  I<X, Y>       --> LHS               RHS

f<X, bool>
  I<u64, bool>  --> RHS               LHS
  I<u64, Y>     --> Both(u64, bool)   Both(X, Y)
  I<X, bool>    --> Match             Match
  I<X, Y>       --> LHS               RHS

f<X, Y>
  I<u64, bool>  --> RHS               LHS
  I<u64, Y>     --> RHS               LHS
  I<X, bool>    --> RHS               LHS
  I<X, Y>       --> Match             Match

In the old behavior (and in current simulation), only the ones marked
with `Match` are checked (i.e., I is checked in f) but in theory, all
pairs should be checked.