[executor] keep root SMT tree as cache
This commit changes the prune strategy of SMT. Before the change, all committed tree are pruned immediately,
after the change, we only prune the previous root state tree when update to new root block.

This effectively delays the root SMT pruning and can increase the hit rate when it's served as base view for execution.